moved api logic to new file
